+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools">
 
     <uses-permission android:name="android.permission.INTERNET" />
+    <uses-permission android:name="android.permission.ACCESS_WIFI_STATE"/>
 
     <application
         android:allowBackup="true"
@@ -16,9 +17,6 @@
         tools:targetApi="31">
         <activity
             android:name=".DeveloperListActivity"
-            android:exported="false" />
-        <activity
-            android:name=".DevelopersList"
             android:exported="false" />
         <activity
             android:name=".About_Us"
